requirements.txt
fastapi
uvicorn[standard]
pydantic
google-adk
python-dotenv
requests
beautifulsoup4
email-validator
